[FIx]
sendRemoveUserFromTaskEvent 수정
